requirements.txt
tensorflow-macos
tensorflow-metal
numpy
matplotlib
pillow
tqdm
scikit-learn